Remove uniqueness constraint from username fields
Ideally, we would, in fact, want UserProfile.unix_username and sql_username to be unique allowing NULL.  However, Django interprets a blank form field as the empty string, not as NULL, and so if the uniqueness constraint were enabled, only one user would be allowed to have a blank username.  Thus, we're temporarily disabling uniqueness.